test(AppHeader): add image logo story and viewports
