+++ src/components/AppHeader/AppHeader.stories.tsx
@@ -1,0 +1,143 @@
+import type { StoryObj, Meta } from '@storybook/react';
+import React from 'react';
+
+import { AppHeader } from './AppHeader';
+import { chromaticViewports } from '../../util/viewports';
+
+export default {
+  title: 'Components/AppHeader',
+  component: AppHeader,
+  parameters: {
+    badges: ['api-1.0', 'theme-2.0'],
+    chromatic: {
+      viewports: [
+        chromaticViewports.googlePixel2,
+        chromaticViewports.ipadMini,
+        chromaticViewports.chromebook,
+      ],
+    },
+  },
+} as Meta<typeof AppHeader>;
+
+type Story = StoryObj<typeof AppHeader>;
+
+export const Default: Story = {
+  args: {
+    title: 'Title',
+    subTitle: 'Sub-title',
+    navGroups: [
+      {
+        name: 'group-1',
+        navItems: [
+          {
+            name: 'Link Item One',
+            type: 'link',
+            href: 'https://example.org',
+          },
+          {
+            name: 'Link Item Two',
+            type: 'link',
+            href: 'https://example.org',
+          },
+          {
+            name: 'Link Item Three',
+            type: 'link',
+            href: 'https://example.org',
+          },
+        ],
+      },
+      {
+        name: 'group-2',
+        navItems: [
+          {
+            name: 'Sign out',
+            type: 'link',
+            href: 'https://example.org',
+            icon: 'logout',
+            iconLayout: 'left',
+          },
+        ],
+      },
+    ],
+  },
+};
+
+export const Buttons: Story = {
+  args: {
+    title: 'Title',
+    subTitle: 'Sub-title',
+    navGroups: [
+      {
+        name: 'group-1',
+        navItems: [
+          {
+            name: 'Button Item One',
+            type: 'button',
+          },
+          {
+            name: 'Button Item Two',
+            type: 'button',
+          },
+          {
+            name: 'sep1',
+            type: 'separator',
+          },
+          {
+            name: 'Button Item Three',
+            type: 'link',
+            href: 'https://example.org',
+          },
+        ],
+      },
+      {
+        name: 'group-2',
+        navItems: [
+          {
+            name: 'Sign out',
+            type: 'link',
+            href: 'https://example.org',
+            icon: 'logout',
+            iconLayout: 'left',
+          },
+        ],
+      },
+    ],
+  },
+};
+
+export const VerticalOrientation: Story = {
+  args: {
+    ...Buttons.args,
+    orientation: 'vertical',
+  },
+};
+
+export const ImageLogo: Story = {
+  args: {
+    ...Default.args,
+    title: (
+      <svg
+        fill="none"
+        height="46px"
+        viewBox="0 0 158 46"
+        width="158px"
+        xmlns="http://www.w3.org/2000/svg"
+      >
+        <path
+          d="M51 7.23041H55.9027V6.0068H52.4779V0H51V7.23041ZM64.3585 9C65.1331 9 65.8058 8.81798 66.2441 8.41348C66.6314 8.05954 66.8659 7.56403 66.8659 6.83593V2.04263H65.5306V2.5887H65.5102C65.1943 2.15386 64.7152 1.90105 64.0629 1.90105C62.7378 1.90105 61.8001 2.89207 61.8001 4.43928C61.8001 6.00671 62.9417 6.91683 64.1036 6.91683C64.7662 6.91683 65.1637 6.65391 65.4695 6.31008H65.5V6.87638C65.5 7.58426 65.1229 7.95842 64.3381 7.95842C63.6959 7.95842 63.4003 7.7056 63.2882 7.38201H61.9122C62.0549 8.39325 62.9315 9 64.3585 9ZM64.3381 5.81457C63.6246 5.81457 63.1557 5.29884 63.1557 4.41906C63.1557 3.54938 63.6246 3.00331 64.3279 3.00331C65.1637 3.00331 65.5714 3.65051 65.5714 4.40894C65.5714 5.17749 65.2146 5.81457 64.3381 5.81457ZM73.3314 1.23372H74.7176V0H73.3314V1.23372ZM73.3314 7.23041H74.7176V2.04272H73.3314V7.23041ZM83.274 7.38201C84.5787 7.38201 85.4757 6.75503 85.4757 5.71345C85.4757 4.49996 84.5074 4.25726 83.6308 4.07523C82.8867 3.92354 82.1936 3.88309 82.1936 3.43815C82.1936 3.06399 82.5504 2.86174 83.0906 2.86174C83.6817 2.86174 84.0385 3.06399 84.0996 3.62017H85.3534C85.2514 2.57859 84.487 1.90105 83.111 1.90105C81.9184 1.90105 80.9807 2.43701 80.9807 3.5595C80.9807 4.69209 81.898 4.9449 82.8357 5.12693C83.5492 5.2685 84.2118 5.31906 84.2118 5.81457C84.2118 6.17862 83.8652 6.41121 83.2536 6.41121C82.6319 6.41121 82.2038 6.14829 82.1121 5.55165H80.8278C80.9093 6.65391 81.7553 7.38201 83.274 7.38201ZM91.3555 7.23032H92.7417V4.21681C92.7417 3.52916 93.1188 3.08421 93.6794 3.08421C94.189 3.08421 94.4846 3.38758 94.4846 3.97411V7.23032H95.8708V4.21681C95.8708 3.52916 96.2276 3.08421 96.8086 3.08421C97.3182 3.08421 97.6138 3.38758 97.6138 3.97411V7.23032H99V3.74152C99 2.60893 98.3782 1.90105 97.2774 1.90105C96.6149 1.90105 96.0645 2.24488 95.7077 2.81117H95.6874C95.4325 2.2651 94.9025 1.90105 94.24 1.90105C93.5163 1.90105 93.0067 2.2651 92.7213 2.74039H92.6907V2.04263H91.3555V7.23032ZM90.5988 7.23044V2.04275H89.2126V5.03604C89.2126 5.72369 88.815 6.20908 88.1627 6.20908C87.5715 6.20908 87.2963 5.87537 87.2963 5.26863V2.04275H85.9203V5.50121C85.9203 6.63381 86.5726 7.37202 87.7346 7.37202C88.4685 7.37202 88.8762 7.09898 89.2329 6.62369H89.2635V7.23044H90.5988ZM75.4709 8.92921H76.8571V6.6438H76.8775C77.1731 7.08874 77.6522 7.38201 78.386 7.38201C79.7315 7.38201 80.6488 6.3202 80.6488 4.64153C80.6488 3.02354 79.762 1.90105 78.3758 1.90105C77.6623 1.90105 77.1731 2.23476 76.8367 2.68982H76.8062V2.04263H75.4709V8.92921ZM78.0904 6.2393C77.2648 6.2393 76.8265 5.62244 76.8265 4.68198C76.8265 3.75163 77.1731 3.01342 78.0395 3.01342C78.8957 3.01342 79.2422 3.70107 79.2422 4.68198C79.2422 5.66289 78.7937 6.2393 78.0904 6.2393ZM70.0953 7.38201C71.716 7.38201 72.827 6.18874 72.827 4.64153C72.827 3.09432 71.716 1.90105 70.0953 1.90105C68.4747 1.90105 67.3637 3.09432 67.3637 4.64153C67.3637 6.18874 68.4747 7.38201 70.0953 7.38201ZM70.0953 6.33031C69.2391 6.33031 68.7703 5.65278 68.7703 4.64153C68.7703 3.63028 69.2391 2.94264 70.0953 2.94264C70.9413 2.94264 71.4204 3.63028 71.4204 4.64153C71.4204 5.65278 70.9413 6.33031 70.0953 6.33031ZM58.8188 7.38201C60.4395 7.38201 61.5505 6.18874 61.5505 4.64153C61.5505 3.09432 60.4395 1.90105 58.8188 1.90105C57.1982 1.90105 56.0872 3.09432 56.0872 4.64153C56.0872 6.18874 57.1982 7.38201 58.8188 7.38201ZM58.8188 6.33031C57.9626 6.33031 57.4938 5.65278 57.4938 4.64153C57.4938 3.63028 57.9626 2.94264 58.8188 2.94264C59.6648 2.94264 60.1439 3.63028 60.1439 4.64153C60.1439 5.65278 59.6648 6.33031 58.8188 6.33031Z"
+          fill="#6C798B"
+        ></path>
+        <path
+          d="M52 39.9745H54.3591V35.1586H59.419V33.1411H54.3591V30.3427H60.3301V28.3416H52V39.9745ZM68.7208 40.2184C70.9497 40.2184 72.3327 38.8029 72.4791 37.0295H70.3477C70.2013 37.9731 69.5668 38.445 68.6719 38.445C67.3216 38.445 66.687 37.3874 66.687 35.8092C66.687 34.1985 67.3704 33.1735 68.7045 33.1735C69.518 33.1735 70.0874 33.7267 70.2339 34.4589H72.3977C72.1049 32.6529 70.7708 31.4001 68.6069 31.4001C66.0688 31.4001 64.4418 33.32 64.4418 35.8092C64.4418 38.2985 66.0851 40.2184 68.7208 40.2184ZM78.5751 30.3265H80.7878V28.3416H78.5751V30.3265ZM78.5751 39.9745H80.7878V31.6281H78.5751V39.9745ZM93.4487 31.6279H91.2848V39.9743H93.4975V35.2398C93.4975 34.1009 94.2134 33.32 95.1896 33.32C96.1169 33.32 96.6376 33.9545 96.6376 34.8493V39.9743H98.8503V34.5239C98.8503 32.7017 97.6951 31.4001 95.9705 31.4001C94.8316 31.4001 94.0669 31.872 93.4975 32.7668H93.4487V31.6279ZM108.476 39.9745H110.689V28.3416H108.476V39.9745ZM127.354 40.2184C129.941 40.2184 131.715 38.2985 131.715 35.8092C131.715 33.32 129.941 31.4001 127.354 31.4001C124.767 31.4001 122.994 33.32 122.994 35.8092C122.994 38.2985 124.767 40.2184 127.354 40.2184ZM127.354 38.5263C125.988 38.5263 125.239 37.4362 125.239 35.8092C125.239 34.1823 125.988 33.0759 127.354 33.0759C128.705 33.0759 129.469 34.1823 129.469 35.8092C129.469 37.4362 128.705 38.5263 127.354 38.5263ZM145.53 40.2184C148.117 40.2184 149.89 38.2985 149.89 35.8092C149.89 33.32 148.117 31.4001 145.53 31.4001C142.943 31.4001 141.169 33.32 141.169 35.8092C141.169 38.2985 142.943 40.2184 145.53 40.2184ZM145.53 38.5263C144.163 38.5263 143.415 37.4362 143.415 35.8092C143.415 34.1823 144.163 33.0759 145.53 33.0759C146.88 33.0759 147.645 34.1823 147.645 35.8092C147.645 37.4362 146.88 38.5263 145.53 38.5263ZM52 23.5483H56.3443C57.8087 23.5483 58.9314 23.1478 59.7938 22.4431C61.0141 21.434 61.6487 19.8002 61.6487 17.9423C61.6487 14.5626 59.6473 12.096 56.5396 12.096H52V23.5483ZM54.3593 21.5461V14.0821H56.2467C58.248 14.0821 59.2568 15.6678 59.2568 17.9423C59.2568 20.2167 58.3945 21.5461 56.1816 21.5461H54.3593ZM71.3324 26.2392H73.5452V22.6193H73.5778C74.0496 23.324 74.8143 23.7885 75.9859 23.7885C78.1336 23.7885 79.598 22.1067 79.598 19.4479C79.598 16.8851 78.1824 15.1072 75.9696 15.1072C74.8306 15.1072 74.0496 15.6358 73.5127 16.3566H73.4639V15.3315H71.3324V26.2392ZM75.514 21.9786C74.1961 21.9786 73.4964 21.0016 73.4964 19.512C73.4964 18.0384 74.0496 16.8691 75.4326 16.8691C76.7994 16.8691 77.3526 17.9583 77.3526 19.512C77.3526 21.0656 76.6367 21.9786 75.514 21.9786ZM90.989 15.3315H88.8575V23.5483H91.0703V19.544C91.0703 17.8141 92.1117 17.0613 93.5598 17.2055H93.6086V15.2994C93.4784 15.2514 93.332 15.2354 93.0879 15.2354C92.1117 15.2354 91.5422 15.7159 91.0378 16.6449H90.989V15.3315ZM99.7726 23.5483H101.985V18.7752C101.985 17.686 102.587 16.9812 103.482 16.9812C104.296 16.9812 104.768 17.4618 104.768 18.3908V23.5483H106.981V18.7752C106.981 17.686 107.55 16.9812 108.478 16.9812C109.291 16.9812 109.763 17.4618 109.763 18.3908V23.5483H111.976V18.0224C111.976 16.2284 110.983 15.1072 109.226 15.1072C108.168 15.1072 107.29 15.6518 106.72 16.5488H106.688C106.281 15.6839 105.435 15.1072 104.377 15.1072C103.222 15.1072 102.408 15.6839 101.953 16.4367H101.904V15.3315H99.7726V23.5483ZM124.086 15.3315H121.922V23.5483H124.134V18.8873C124.134 17.7661 124.85 16.9973 125.827 16.9973C126.754 16.9973 127.275 17.6219 127.275 18.5029V23.5483H129.488V18.1825C129.488 16.3886 128.332 15.1072 126.608 15.1072C125.469 15.1072 124.704 15.5717 124.134 16.4527H124.086V15.3315ZM141.627 23.7725C142.994 23.7725 144.073 23.276 144.865 22.2829C145.657 21.2899 146.052 20.0993 146.052 18.7111C146.052 17.6006 145.722 16.7143 145.06 16.0522C144.398 15.3902 143.552 15.0592 142.522 15.0592C141.187 15.0592 140.114 15.561 139.3 16.5648C138.486 17.5579 138.08 18.7538 138.08 20.1526C138.08 21.2845 138.41 22.1708 139.072 22.8115C139.745 23.4522 140.596 23.7725 141.627 23.7725ZM139.381 20.1366C139.381 19.0368 139.653 18.0918 140.195 17.3016C140.748 16.5114 141.513 16.1163 142.489 16.1163C143.227 16.1163 143.785 16.3619 144.165 16.8531C144.545 17.3443 144.734 17.9636 144.734 18.7111C144.734 19.8323 144.463 20.7773 143.921 21.5461C143.378 22.315 142.63 22.6994 141.675 22.6994C140.938 22.6994 140.368 22.4644 139.967 21.9946C139.577 21.5141 139.381 20.8948 139.381 20.1366ZM146.935 16.3086H148.074L146.706 23.3063L146.659 23.5484H147.944L149.312 16.5507L149.36 16.3086H150.906L151.117 15.2675H149.571L149.815 14.0022C149.88 13.6818 149.989 13.4522 150.141 13.3134C150.293 13.1746 150.569 13.1052 150.971 13.1052C151.155 13.1052 151.323 13.1105 151.475 13.1212L151.687 12.032C151.448 12.0107 151.209 12 150.971 12C150.255 12 149.707 12.1495 149.327 12.4485C148.948 12.7475 148.687 13.244 148.546 13.9381L148.286 15.2675H147.147L146.935 16.3086ZM130.069 16.773H131.224V21.5942C131.224 23.1479 132.412 23.5643 133.486 23.5643C134.364 23.5643 134.869 23.5323 134.869 23.5323V21.9306C134.869 21.9306 134.495 21.9466 134.218 21.9466C133.73 21.9466 133.388 21.7384 133.388 21.1297V16.773H134.82V15.3315H133.388V12.7687H131.224V15.3315H130.069V16.773ZM117.128 23.7885C119.276 23.7885 120.594 22.5552 120.903 21.0977H118.723C118.479 21.7223 117.974 22.1067 117.112 22.1067C115.859 22.1067 115.143 21.3219 114.98 20.0566H121.033C121.033 17.1254 119.601 15.1072 116.916 15.1072C114.476 15.1072 112.8 16.9973 112.8 19.4319C112.8 21.8825 114.362 23.7885 117.128 23.7885ZM116.949 16.7891C117.958 16.7891 118.658 17.5098 118.723 18.5509H115.013C115.208 17.4778 115.777 16.7891 116.949 16.7891ZM94.0176 16.773H95.1728V21.5942C95.1728 23.1479 96.3606 23.5643 97.4345 23.5643C98.3131 23.5643 98.8175 23.5323 98.8175 23.5323V21.9306C98.8175 21.9306 98.4433 21.9466 98.1667 21.9466C97.6786 21.9466 97.3369 21.7384 97.3369 21.1297V16.773H98.7687V15.3315H97.3369V12.7687H95.1728V15.3315H94.0176V16.773ZM85.7477 23.5483H87.8792V23.4683C87.6839 23.3241 87.5863 22.9557 87.5863 22.331V18.1505C87.5863 16.2285 86.5775 15.1233 84.0718 15.1233C81.6311 15.1233 80.541 16.3887 80.4759 17.8302H82.4447C82.5098 17.0454 83.0142 16.677 84.0555 16.677C84.983 16.677 85.4548 17.0614 85.4548 17.654C85.4548 18.2787 84.8202 18.3908 83.4372 18.583C81.5823 18.8393 80.0854 19.4159 80.0854 21.2259C80.0854 22.8596 81.2895 23.7405 82.9979 23.7405C84.4297 23.7405 85.0318 23.26 85.4548 22.6354H85.4874C85.5362 23.0038 85.6175 23.3882 85.7477 23.5483ZM83.5348 22.2189C82.7538 22.2189 82.2332 21.9466 82.2332 21.1778C82.2332 20.409 82.8027 20.1527 83.9579 19.8964C84.5111 19.7683 85.1131 19.6401 85.5036 19.4159V20.6492C85.5036 21.4501 84.804 22.2189 83.5348 22.2189ZM66.5388 23.7885C68.6865 23.7885 70.0045 22.5552 70.3136 21.0977H68.1333C67.8892 21.7223 67.3848 22.1067 66.5225 22.1067C65.2696 22.1067 64.5537 21.3219 64.391 20.0566H70.4438C70.4438 17.1254 69.0119 15.1072 66.3272 15.1072C63.8866 15.1072 62.2107 16.9973 62.2107 19.4319C62.2107 21.8825 63.7727 23.7885 66.5388 23.7885ZM66.3598 16.7891C67.3686 16.7891 68.0682 17.5098 68.1333 18.5509H64.4235C64.6188 17.4778 65.1883 16.7891 66.3598 16.7891ZM154.167 40.2184C156.25 40.2184 157.681 39.2096 157.681 37.5338C157.681 35.5815 156.136 35.191 154.737 34.8981C153.549 34.6541 152.442 34.589 152.442 33.8731C152.442 33.2712 153.012 32.9458 153.874 32.9458C154.818 32.9458 155.387 33.2712 155.485 34.166H157.486C157.323 32.4902 156.103 31.4001 153.907 31.4001C152.003 31.4001 150.506 32.2624 150.506 34.0684C150.506 35.8906 151.971 36.2973 153.467 36.5902C154.606 36.818 155.664 36.8993 155.664 37.6965C155.664 38.2823 155.111 38.6565 154.135 38.6565C153.142 38.6565 152.459 38.2334 152.312 37.2735H150.262C150.392 39.0469 151.743 40.2184 154.167 40.2184ZM136.285 42.8215C137.521 42.8215 138.595 42.5287 139.295 41.8779C139.913 41.3084 140.287 40.5112 140.287 39.3398V31.6279H138.156V32.5065H138.123C137.619 31.8069 136.854 31.4001 135.813 31.4001C133.698 31.4001 132.201 32.9946 132.201 35.4839C132.201 38.0057 134.023 39.47 135.878 39.47C136.935 39.47 137.57 39.0469 138.058 38.4938H138.107V39.4049C138.107 40.5438 137.505 41.1457 136.252 41.1457C135.227 41.1457 134.755 40.739 134.576 40.2184H132.38C132.608 41.8453 134.007 42.8215 136.285 42.8215ZM136.252 37.6965C135.113 37.6965 134.365 36.8668 134.365 35.4513C134.365 34.0521 135.113 33.1735 136.236 33.1735C137.57 33.1735 138.221 34.2148 138.221 35.435C138.221 36.6715 137.651 37.6965 136.252 37.6965ZM114.786 39.9745H122.612V38.0058H117.145V28.3416H114.786V39.9745ZM105.334 39.9743H107.465V39.893C107.27 39.7465 107.172 39.3723 107.172 38.7378V34.4914C107.172 32.539 106.164 31.4164 103.658 31.4164C101.218 31.4164 100.127 32.7017 100.062 34.166H102.031C102.096 33.3688 102.6 32.9946 103.642 32.9946C104.569 32.9946 105.041 33.3851 105.041 33.987C105.041 34.6216 104.406 34.7354 103.023 34.9307C101.169 35.191 99.6719 35.7767 99.6719 37.6152C99.6719 39.2747 100.876 40.1696 102.584 40.1696C104.016 40.1696 104.618 39.6815 105.041 39.0469H105.073C105.122 39.4211 105.204 39.8116 105.334 39.9743ZM103.121 38.6239C102.34 38.6239 101.819 38.3473 101.819 37.5664C101.819 36.7854 102.389 36.5251 103.544 36.2648C104.097 36.1346 104.699 36.0045 105.09 35.7767V37.0295C105.09 37.843 104.39 38.6239 103.121 38.6239ZM86.0316 40.2184C88.6185 40.2184 90.3919 38.2985 90.3919 35.8092C90.3919 33.32 88.6185 31.4001 86.0316 31.4001C83.4447 31.4001 81.6713 33.32 81.6713 35.8092C81.6713 38.2985 83.4447 40.2184 86.0316 40.2184ZM86.0316 38.5263C84.6649 38.5263 83.9165 37.4362 83.9165 35.8092C83.9165 34.1823 84.6649 33.0759 86.0316 33.0759C87.382 33.0759 88.1467 34.1823 88.1467 35.8092C88.1467 37.4362 87.382 38.5263 86.0316 38.5263ZM72.8205 33.0922H73.9756V37.9894C73.9756 39.5676 75.1633 39.9906 76.2372 39.9906C77.1157 39.9906 77.6201 39.9581 77.6201 39.9581V38.3311C77.6201 38.3311 77.2459 38.3474 76.9693 38.3474C76.4812 38.3474 76.1395 38.1358 76.1395 37.5176V33.0922H77.5713V31.6279H76.1395V29.0248H73.9756V31.6279H72.8205V33.0922ZM61.3457 39.9745H63.5584V31.6281H61.3457V39.9745ZM61.3457 30.3265H63.5584V28.3416H61.3457V30.3265Z"
+          fill="#212F43"
+        ></path>
+        <path
+          d="M15 29.031V36.031H4V0.0310059H0V40.031H15V45.031H27.5C34.4036 45.031 40 39.4346 40 32.531C40 30.0315 39.2637 27.705 38 25.7517V10.031H22V14.031H35V20.031H22V29.031H26V24.031H27.5C32.1944 24.031 36 27.8366 36 32.531C36 37.2254 32.1944 41.031 27.5 41.031H19V33.031H22V29.031H15V4.03101H38V0.0310059H11V29.031H15Z"
+          fill="#1E79F8"
+        ></path>
+      </svg>
+    ),
+    href: 'https://example.org',
+    subTitle: undefined,
+  },
+};
